pyproject.toml
[tool.poetry]
name = "poetry-calendar-api"
version = "0.1.0"
description = ""
authors = ["Ignacio gallo <ignacioga.346@gmail.com>"]
readme = "README.md"
packages = [{include = "poetry_calendar_api"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
asgiref = "^3.5.2"
django-cors-headers = "^3.13.0"
djangorestframework-simplejwt = "^5.2.0"
PyJWT = "^2.5.0"
pytz = "^2022.2.1"
sqlparse = "^0.4.2"
tzdata = "^2022.2"
dj-database-url = "^1.0.0"
psycopg2-binary = "^2.9.3"
whitenoise = {extras = ["brotli"], version = "^6.2.0"}
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
